fastapi/requirements.txt
openai
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
aiomysql
pymysql
sqlparse